logger updates for enum/interfaces
