+++ components/amorphic-contracts/index.d.ts
@@ -1,3 +1,6 @@
+import { Enums } from './logger-enums';
+import { Interfaces } from './logger-interfaces';
+
 /**
  * application configuration store object
  *
@@ -28,142 +31,110 @@ export interface StatsdClientInterface {
     counter(statsKey: string, num: number, tags?: object): void
 }
 
-/**
- * Logger Object types and interfaces
- */
- export namespace Logger {
-    export type Data = {
-        [key: string]: any;
-    }
+export interface LoggerCommon {
+    /**
+     * This utility function converts the errData to ErrorLog that can be directly used with logger.error().
+     * It adds isHumanRelated and code fields to the result from errData.
+     * It adds message to the result from error or errData when possible. Use "error" when message not provided.
+     * It adds name and stack fields to result when the fields are not empty.
+     * If the error contains fields other than name, stack, and message, it will put them into
+     * logErrorObj.data.fromError.
+     * @param {Interfaces.ErrorData} errData - error data with mandatory fields and optional message.
+     * @returns {Interfaces.ErrorLog} - returns logErrorObj
+     */
+    mapToErrorLog?(errData: Interfaces.ErrorData): Interfaces.ErrorLog;
+}
 
-    export type strNumKeyValuePair = {
-        [key: string]: number | string;
-    }
+export interface Logger extends Interfaces.BunyanBaseInterface, LoggerCommon {
 
-    export interface ErrorData {
-        isHumanRelated: boolean;
-        code?: string;
-        error?: any;
-        message?: string;
-        data?: Data & { fromError?: never };
-    }
+    /**
+     * Uses the request header's groupId and source value to set local storage context object.
+     * By optionally setting 'generateContextIfMissing' to true, the middleware also has the ability to
+     * generate new globalId and globalSource header values if the current values are missing in
+     * request. Default value for 'generateContextIfMissing' is false.
+     * @param {Interfaces.ApiContextMiddleware} options - This is an optional field. generateContextIfMissing
+     * can be optionally specified as true (default is false), to create new global id and global source.
+     * Note: The new globalId and globalSource would only be created if they are both not populated by request.
+     * @returns {any} - returns any
+     */
+    setApiContextMiddleware?(options?: Interfaces.ApiContextMiddleware) : any;
 
-    export interface Context {
-        globalId?: string,
-        globalSource?: string,
-        referenceId?: string,
-        referenceType?: string,
-        sessionId? : string,
-        piiLevel?: any,
-        data?: Data
-    }
+    /**
+     * This utility function would fetch the headers info from the current local storage context.
+     * This can be useful in cases when you may want to propogate exisiting cached context header
+     * to a new service.
+     * @returns {Enums.GlobalHeaderType} - returns header object or type GlobalHeaderType
+     */
+    getHeaderContextFromLocalStorage?(): Enums.GlobalHeaderType;
 
-    export interface ErrorLog {
-        isHumanRelated?: boolean,
-        code?: string,
-        name?: string,
-        message?: string,
-        stack?: string,
-        data?: Data
-    }
+    /**
+     * Fetches current context from LocalStorage. This function returns undefined if called outside of a
+     * asynchronous context initialized by calling "asyncLocalStorage.run"
+     * @returns {Interfaces.Context} - returns Logger Context object
+     */
+    getContextFromLocalStorage?(): Interfaces.Context;
 
-    interface GeoIP  {
-        latitude?: number,
-        longitude?: number,
-        cityName?: string,
-        stateName?: string,
-        postalCode?: string,
-        timeZone?: string,
-        countryName?: string,
-        countryCode?: string,
-        continentCode?: string
-    }
+    /**
+     * Uses the context object to set the entire local storage context (not just the headers).
+     * By optionally setting 'generateContextIfMissing' to true, the middleware also has the ability to
+     * generate new globalId and globalSource header values if the current values are missing in
+     * request. Default value for 'generateContextIfMissing' is false.
+     * @param {Interfaces.SourceContextMiddleware} options - context is required here. generateContextIfMissing
+     * can be optionally specified as true (default is false), to create new global id and source.
+     * @param {any} next - next items to run
+     * @returns {Promise<any>} - returns promise
+     */
+    setContextFromSourceMiddleware?(options: Interfaces.SourceContextMiddleware, next): Promise<any>;
 
-    export interface Request {
-        url?: string,
-        method?: string,
-        clientIpAddress?: string,
-        data?: Data,
-        remoteUser?: string,
-        remoteAddress?: string,
-        line?: string,
-        length?: number,
-        time?: number,
-        headers?: string,
-        status?: number,
-        bytesSent?: number,
-        bodyBytesSent?: number,
-        httpReferer?: string,
-        httpUserAgent?: string,
-        httpXForwardedFor?: string,
-        upstreamAddress?: string,
-        upstreamStatus?: number,
-        upstreamResponseTime?: number,
-        upstreamConnectTime?: number,
-        upstreamHeaderTime?: number,
-        upstreamBytesReceived?: number,
-        upstreamBytesSent?: number,
-        geoip?: GeoIP
-    }
+    /**
+     * Set endpoint GET /uiLogConfig on backend server for UI Logger to fetch backend logLevel.
+     * @param {Enums.ServerType} serverType - supported server type (express or restify) is required.
+     * @param {any} server - server instance (express or restify) is required.
+     * @returns void
+     */
+    addUIConfigHandler?(serverType: Enums.ServerType, server: any): void;
 
-    export interface Response {
-        status?: number,
-        data?: Data
-    }
-    
-    export interface Log {
-        module?: string,
-        function?: string,
-        category?: 'security' | 'availability' | 'request' | 'milestone'
-        message?: string,
-        context?: Context,
-        error?: ErrorLog,
-        request?: Request,
-        response?: Response,
-        data?: Data
-    }    
+    /**
+     * Set endpoint POST /uiLog on backend server for parsing logs sent from UI Logger.
+     * Backend server would use this api to receive UI logs and use its logger to parse UI logs.
+     * @param {Enums.ServerType} serverType - supported server type (express or restify) is required.
+     * @param {any} server - server instance (express or restify) is required.
+     * @param {Logger} logger - logger instance (express or restify) is required.
+     * @returns void
+     */
+    addUILogHandler?(serverType: Enums.ServerType, server: any, logger: Logger): void;
 
-    export enum PiiLevelEnum {
-        's1' = 's1',
-        's2' = 's2',
-        'nonPII' = 'non-pii'
-    }    
+    /**
+     * Use this method to optionally set defaults for module, category, context.piiLevel,
+     * error.isHumanRelated in the first parameter as an object. You may also use the
+     * second parameter of the function to pass key/value pairs. These key/value pairs would
+     * get set as default data elements on the log. Defaults set by this function can always be overwritten
+     * by the individual log. For data element we merge the default's data element with log's data
+     * element.
+     * @param {Interfaces.ChildLogRootValues} rootValues - Optional. set module, category, context.piiLevel,
+     * error.isHumanRelated defaults.
+     * @param {Interfaces.ChildLogDataValues} dataValues - Optional. set key/value pair defaults, that get added
+     * to the root level data objects.
+     * @returns {Logger}
+     */
+    childLogger?(rootValues?: Interfaces.ChildLogRootValues, dataValues?: Interfaces.ChildLogDataValues): Logger;
 
-    export enum Category {
-        security = 'security',
-        availability = 'availability',
-        request = 'request',
-        milestone = 'milestone'
-    }
+}
 
-    export interface BunyanBaseInterface {
-        info(): boolean;
-        error(): boolean;
-        debug(): boolean;
-        warn(): boolean;
-        info(object: ExpandedLog, ...params: any[]): void;
-        info(message: string, functionName: string, piiLevel?: PiiLevelEnum): void;
-        error(object: ExpandedLog, ...params: any[]): void;
-        debug(object: ExpandedLog, ...params: any[]): void;
-        warn(object: ExpandedLog, ...params: any[]): void;
-    }
+export interface UILogger extends Interfaces.BunyanBaseInterface, LoggerCommon {
 
-    export interface ChildLogRootValues {
-        module?: string,
-        category?: Category,
-        context?: {
-            piiLevel?: PiiLevelEnum
-        },
-        error?:  {
-            isHumanRelated?: boolean
-        }
-    }
-    
-    export interface ChildLogDataValues {
-        [key: string]: any;
-    }
-    
-    export interface ExpandedLog extends Log {
-        [key: string]: any;
-    }
+    /**
+     * Use this method to optionally set defaults for module, category, context.piiLevel,
+     * error.isHumanRelated in the first parameter as an object. You may also use the
+     * second parameter of the function to pass key/value pairs. These key/value pairs would
+     * get set as default data elements on the log. Defaults set by this function can always be overwritten
+     * by the individual log. For data element however, we merge the default's data element with log's data
+     * element.
+     * @param {Interfaces.ChildLogRootValues} rootValues - Optional. set module, category, context.piiLevel,
+     * error.isHumanRelated defaults.
+     * @param {Interfaces.ChildLogDataValues} dataValues - Optional. set key/value pair defaults, that get added
+     * to the root level data objects.
+     * @returns {UILogger}
+     */
+    childLogger?(rootValues?: Interfaces.ChildLogRootValues, dataValues?: Interfaces.ChildLogDataValues): UILogger;
 }
